String to Huffman code
This is the implement of transform a string to Huffman code. This is helpful to Task4 because in Task4, we need to input the name of a textfile instead  of a string. So if we can change the program to make it accept a text file name, that would be great.